+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 Changes in 0.6.X
 ================
 
+- Fixed queryset manager issue (MongoEngine/mongoengine#52)
 - Fixed FileField comparision (hmarr/mongoengine#547)
 
 Changes in 0.6.18